Update sql schema script

+++ dispatch-controller-service/src/main/resources/schema.sql
@@ -2,8 +2,8 @@ CREATE TABLE drone (
     id INTEGER NOT NULL AUTO_INCREMENT,
     serial_number VARCHAR(100) NOT NULL UNIQUE,
     model ENUM('Lightweight', 'Middleweight', 'Cruiserweight', 'Heavyweight') NOT NULL,
-    weight_limit DECIMAL(5, 2) NOT NULL,
-    battery_capacity DECIMAL(3, 2) NOT NULL,
+    weight_limit DOUBLE NOT NULL,
+    battery_capacity DOUBLE NOT NULL,
     state ENUM('IDLE', 'LOADING', 'LOADED', 'DELIVERING', 'DELIVERED', 'RETURNING') NOT NULL,
     PRIMARY KEY (id)
 );
@@ -21,7 +21,7 @@ CREATE TABLE medication (
     shipment_id INTEGER NOT NULL,
     code VARCHAR(100) NOT NULL,
     name VARCHAR(100) NOT NULL,
-    weight_limit DECIMAL(5, 2) NOT NULL,
+    weight_limit DOUBLE NOT NULL,
     image VARCHAR(100) NOT NULL,
     PRIMARY KEY (id),
     FOREIGN KEY (shipment_id) REFERENCES shipment(id)
